cloud_functions/requirements.txt
functions-framework==3.*
google-cloud-storage
joblib
scikit-learn